+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [v0.9.0](https://github.com/ameily/cincoconfig/releases/tag/v0.9.0) - 2023-04-23
 ### Added
 - New support method, `is_value_defined` to check if a config field value is defined by the user,
   through either a loaded config file or the API.